Revert "fixed data types"
This reverts commit e586e2c

+++ src/main/java/ru/vsu/noidle_server/model/domain/AchievementEntity.java
@@ -8,7 +8,6 @@ import ru.vsu.noidle_server.model.converter.SubTypeConverter;
 import ru.vsu.noidle_server.model.converter.TypeConverter;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
@@ -39,7 +38,7 @@ public class AchievementEntity {
     private String name;
 
     @Column(name = "value", nullable = false)
-    private BigInteger value;
+    private Long value;
 
     @Column(name = "date", nullable = false)
     private Long date;
